# apps/api/docs/swagger/openapi.bundle.yaml
openapi: 3.0.0
info:
  title: Hexagonal NestJS API
  version: 1.0.0
  description: |
    Hexagonal Architecture NestJS + Express API 文件
servers:
  - url: 'http://localhost:3000/api'
    description: 本地開發環境
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        使用 Access Token 進行身份驗證

        格式: Bearer {accessToken}
  schemas:
    SuccessResponse:
      type: object
      description: 通用成功回應格式
      required:
        - success
        - timestamp
      properties:
        success:
          type: boolean
          description: 操作是否成功
          example: true
        data:
          type: object
          description: 回應資料
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: 回應時間
          example: '2024-01-01T00:00:00.000Z'
    ErrorResponse:
      type: object
      description: 通用錯誤回應格式
      required:
        - success
        - timestamp
      properties:
        success:
          type: boolean
          description: 操作是否成功
          example: false
        message:
          type: string
          description: 錯誤訊息
          example: 發生錯誤
        code:
          type: string
          description: 錯誤代碼
          example: ERROR_CODE
        timestamp:
          type: string
          format: date-time
          description: 回應時間
          example: '2024-01-01T00:00:00.000Z'
  responses:
    BadRequest:
      description: 請求參數錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 請求參數錯誤
            code: BAD_REQUEST
            timestamp: '2024-01-01T00:00:00.000Z'
    NoToken:
      description: 缺少授權憑證
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 缺少授權憑證，請先登入
            code: NO_TOKEN
            timestamp: '2024-01-01T00:00:00.000Z'
...
    Forbidden:
      description: 權限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 權限不足
            code: FORBIDDEN
            timestamp: '2024-01-01T00:00:00.000Z'
    NotFound:
      description: 資源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 資源不存在
            code: NOT_FOUND
            timestamp: '2024-01-01T00:00:00.000Z'
...
    InternalServerError:
      description: 伺服器內部錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 伺服器發生未預期錯誤
            code: INTERNAL_SERVER_ERROR
            timestamp: '2024-01-01T00:00:00.000Z'
paths:
...
  /security/ip-whitelist:
    get:
      tags:
        - Security
      summary: 查詢 IP 白名單
      description: 取得所有 IP 白名單記錄（僅限 ADMIN）。
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 查詢成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  - ip: 192.168.1.1
                    description: 辦公室 IP
                    createdAt: '2024-01-01T00:00:00.000Z'
                timestamp: '2024-01-01T00:00:00.000Z'
        '401':
          $ref: '#/components/responses/NoToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Security
      summary: 新增 IP 至白名單
      description: 將指定 IP 加入白名單（僅限 ADMIN）。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ip
              properties:
                ip:
                  type: string
                  description: IP 位址
                  example: 192.168.1.1
                description:
                  type: string
                  description: 備註說明
                  example: 辦公室 IP
      responses:
        '201':
          description: 新增成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  message: IP 192.168.1.1 已加入白名單
                timestamp: '2024-01-01T00:00:00.000Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NoToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/security/ip-whitelist/{ip}':
    delete:
      tags:
        - Security
      summary: 移除 IP 白名單
      description: 將指定 IP 從白名單移除（僅限 ADMIN）。
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: ip
          required: true
          schema:
            type: string
          description: 要移除的 IP 位址
          example: 192.168.1.1
      responses:
        '204':
          description: 移除成功（無回應內容）
        '401':
          $ref: '#/components/responses/NoToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /security/ip-blacklist:
    get:
      tags:
        - Security
      summary: 查詢 IP 黑名單
      description: 取得所有 IP 黑名單記錄（僅限 ADMIN）。
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 查詢成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  - ip: 10.0.0.1
                    reason: 異常登入嘗試
                    createdAt: '2024-01-01T00:00:00.000Z'
                timestamp: '2024-01-01T00:00:00.000Z'
        '401':
          $ref: '#/components/responses/NoToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Security
      summary: 新增 IP 至黑名單
      description: 將指定 IP 加入黑名單（僅限 ADMIN）。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ip
              properties:
                ip:
                  type: string
                  description: IP 位址
                  example: 10.0.0.1
                reason:
                  type: string
                  description: 封鎖原因
                  example: 異常登入嘗試
      responses:
        '201':
          description: 新增成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  message: IP 10.0.0.1 已加入黑名單
                timestamp: '2024-01-01T00:00:00.000Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NoToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/security/ip-blacklist/{ip}':
    delete:
      tags:
        - Security
      summary: 移除 IP 黑名單
      description: 將指定 IP 從黑名單移除（僅限 ADMIN）。
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: ip
          required: true
          schema:
            type: string
          description: 要移除的 IP 位址
          example: 10.0.0.1
      responses:
        '204':
          description: 移除成功（無回應內容）
        '401':
          $ref: '#/components/responses/NoToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /security/unlock-account:
    post:
      tags:
        - Security
      summary: 解鎖帳號
      description: 解除因登入失敗次數超過閾值而被鎖定的帳號（僅限 ADMIN）。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  description: 要解鎖的帳號 Email
                  example: user@example.com
      responses:
        '200':
          description: 解鎖成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  message: 帳號 user@example.com 已解鎖
                timestamp: '2024-01-01T00:00:00.000Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NoToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'